user-specific calendar events and auth to support it

+++ back/Controllers/CalendarController.cs
@@ -1,24 +1,52 @@
 using Microsoft.AspNetCore.Mvc;
 using back.Models;
 using Microsoft.AspNetCore.Authorization;
+using System.Security.Claims;
 
 [ApiController]
 [Route("api/[controller]/[action]")]
-public class CalendarController(TimeManagerDBContext dbContext) : ControllerBase
+public class CalendarController(ILogger<CalendarController> logger, TimeManagerDBContext dbContext) : ControllerBase
 {
+    private bool TryGetUserId(out string userId)
+    {
+        var sub = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier);
+        userId = String.Empty;
+
+        if (sub is null || sub.Value is null)
+        {
+            logger.LogDebug("Request doesn't contain an access token with a valid subject value.");
+
+            return false;
+        }
+
+        userId = sub.Value;
+
+        return true;
+    }
+
     [HttpGet]
     [Authorize]
     public ActionResult<IEnumerable<CalendarEventModel>> Events()
     {
-        var events = dbContext.CalendarEvents.ToArray();
+        if (!TryGetUserId(out var userId))
+        {
+            return new ForbidResult();
+        }
 
-        return new JsonResult(events);
+        var userEvents = dbContext.CalendarEvents.Where(cEvent => cEvent.UserId == userId);
+
+        return new JsonResult(userEvents);
     }
 
     [HttpPost]
     [Authorize]
     public ActionResult<CalendarEventModel> AddEvent([FromBody] NewEventDto newEventDto)
     {
+        if (!TryGetUserId(out var userId))
+        {
+            return new ForbidResult();
+        }
+
         if (!ModelState.IsValid)
         {
             return new BadRequestResult();
@@ -30,6 +58,7 @@ public class CalendarController(TimeManagerDBContext dbContext) : ControllerBase
         var newEvent = new CalendarEventModel()
         {
             Id = Guid.CreateVersion7(),
+            UserId = userId,
             Title = newEventDto.Title,
             StartDateTime = startDateTime,
             EndDateTime = endDateTime
@@ -52,6 +81,11 @@ public class CalendarController(TimeManagerDBContext dbContext) : ControllerBase
     [Authorize]
     public IActionResult EditEvent([FromBody] NewEventDto newEventDto, Guid id)
     {
+        if (!TryGetUserId(out var userId))
+        {
+            return new ForbidResult();
+        }
+
         if (!ModelState.IsValid)
         {
             Console.WriteLine($"Error count: {ModelState.ErrorCount}");
@@ -76,6 +110,11 @@ public class CalendarController(TimeManagerDBContext dbContext) : ControllerBase
             return new NotFoundResult();
         }
 
+        if (eventToEdit.UserId != userId)
+        {
+            return new ForbidResult();
+        }
+
         var newStartDateTime = DateTime.SpecifyKind(newEventDto.StartDateTime, DateTimeKind.Utc);
         var newEndDateTime = DateTime.SpecifyKind(newEventDto.EndDateTime, DateTimeKind.Utc);
 
@@ -92,11 +131,21 @@ public class CalendarController(TimeManagerDBContext dbContext) : ControllerBase
     [Authorize]
     public IActionResult DeleteEvent(Guid id)
     {
+        if (!TryGetUserId(out var userId))
+        {
+            return new ForbidResult();
+        }
+
         var eventToDelete = dbContext.CalendarEvents.Find(id);
 
         if (eventToDelete is null)
         {
             return new NotFoundResult();
+        }
+
+        if (eventToDelete.UserId != userId)
+        {
+            return new ForbidResult();
         }
 
         dbContext.CalendarEvents.Remove(eventToDelete);
